requirements.txt
argcomplete==1.8.2
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.5.3
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cssselect==1.0.3
EbookLib==0.15
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.1
Flask-Cors==3.0.8
geograpy3==0.1.2
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
importlib-metadata==0.19
itsdangerous==1.1.0
jellyfish==0.7.2
jieba3k==0.35.1
Jinja2==2.10.1
lxml==4.4.0
MarkupSafe==1.1.1
more-itertools==7.2.0
newspaper3k==0.2.8
nltk==3.4.4
numpy==1.17.0
packaging==19.1
pandas==0.25.0
Pillow==6.1.0
pluggy==0.12.0
ply==3.11
py==1.8.0
pycountry==19.7.15
pyparsing==2.4.2
PyPDF3==1.0.1
pytest==5.0.1
python-dateutil==2.8.0
python-docx==0.8.10
python-pptx==0.6.5
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-file==1.4.3
six==1.10.0
slate3k==0.5.3
soupsieve==1.9.2
SpeechRecognition==3.6.3
tinysegmenter==0.3
tldextract==2.2.1
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
xlrd==1.0.0
XlsxWriter==1.1.8
zipp==0.5.2
